+++ CHANGELOG.md
@@ -25,6 +25,8 @@ v2.11 (Anoia) - (in development)
         - getFileFilters
         - getDefaultExtension
         - canLoadExtension
+    - The GL plugin interface has changed, so if you have your own GL plugins, you will need to update them.
+      - The interface name changed from `ccGLFilterPluginInterface` to `ccGLPluginInterface`.
 
 v2.10.3 (Zephyrus) - (in development)
 ----------------------